+++ tests-manual/update_test_data.py
@@ -34,6 +34,7 @@ class UpdateData:
         opts: Dict = {},
         verbose: bool = False,
         dry_run: bool = True,
+        full_diff: bool = False,
     ):
         self.dir = dir
         self.globus_token = globus_token
@@ -42,6 +43,7 @@ class UpdateData:
         self.verbose = verbose
         self.upload_verbose = True if "plugin-tests" in dir else False
         self.dry_run = dry_run
+        self.full_diff = full_diff
 
     def log(self, verbose_message, short_message: Optional[str] = None):
         if self.verbose:
@@ -58,12 +60,16 @@ class UpdateData:
             **self.opts,
             verbose=self.upload_verbose,
         )
-        info = self.upload.get_info()
         errors = self.upload.get_errors()
+        info = self.upload.get_info()
         report = ErrorReport(info=info, errors=errors)
-        if "Too many requests" in report.as_md():
+        if "Too Many Requests" in report.as_md():
             raise Exception(
                 f"Something went wrong with Spreadsheet Validator request for {self.dir}."
+            )
+        elif "Unauthorized" in report.as_md():
+            raise Exception(
+                f"URL checking returned 'Unauthorized' in response while checking {self.dir}; did you forget a Globus token?"
             )
         if "fixtures" not in self.exclude:
             new_data = self.update_fixtures(report)
@@ -89,10 +95,17 @@ class UpdateData:
                     json.dump(new_data, f)
         else:
             print(f"{self.dir}/fixtures.json excluded, not changed.")
+        cleaned_report = clean_report(report)
         if "README" not in self.exclude:
             readme = self.open_or_create_readme()
             try:
-                diff_test(self.dir, readme, clean_report(report), verbose=self.verbose)
+                diff_test(
+                    self.dir,
+                    readme,
+                    cleaned_report,
+                    verbose=self.verbose,
+                    full_diff=self.full_diff,
+                )
                 readme.close()
                 print(f"No diff found, skipping {self.dir}/README.md")
             except MockException:
@@ -103,7 +116,7 @@ class UpdateData:
                     self.log(
                         f"""
                             Would have written the following report to {self.dir}/README.md:
-                            {clean_report(report)}
+                            {cleaned_report}
                             """,
                         f"Would have updated {self.dir}/README.md.",
                     )
@@ -112,14 +125,13 @@ class UpdateData:
                     self.log(
                         f"""
                             Writing the following report to {self.dir}/README.md:
-                            {clean_report(report)}
+                            {cleaned_report}
                             """,
                         f"Updating {self.dir}/README.md.",
                     )
                     with open(f"{self.dir}/README.md", "w") as f:
-                        # TODO: potential issues with blank lines at end of report
-                        f.write(clean_report(report))
-                    dataset_test(self.dir, self.opts)
+                        f.write(cleaned_report)
+                    dataset_test(self.dir, self.opts | {"globus_token": self.globus_token})
         else:
             print(f"{self.dir}/README.md excluded, not changed.")
         return self.change_report
@@ -199,7 +211,7 @@ def manual_test(test_dir: Union[str, List], verbose: bool = False):
         assert Path(
             test_dir
         ).resolve(), f"Arg {test_dir} passed to manual_test is not a directory!"
-    elif type(test_dir) is list:
+    elif type(test_dir) is list and len(test_dir) > 1:
         test_dir = [dir for dir in test_dir if Path(dir).is_dir()]
     test = TestDatasetExamples()
     setattr(test, "dataset_test_dirs", test_dir)
@@ -227,6 +239,7 @@ def call_update(dir: str, args) -> Dict:
         dry_run=args.dry_run,
         verbose=args.verbose,
         exclude=args.exclude,
+        full_diff=args.full_diff,
     ).update_test_data()
     return change_report
 
@@ -276,6 +289,12 @@ parser.add_argument(
     "--manual_test",
     action="store_true",
     help="Default is False. Used for investigating testing failures with more verbose output. Requires passing a test_dir. Pass a blank Globus token as this runs offline.",
+)
+parser.add_argument(
+    "-f",
+    "--full_diff",
+    action="store_true",
+    help="Default is False. Show full and cleaned README diff.",
 )
 
 args = parser.parse_args()