+++ tests/fixtures.py
@@ -1,0 +1,123 @@
+SCATACSEQ_HIGHER_VERSION_VALID = {
+    "test-schema-v0.0": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": True,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+    "test-schema-v0.1": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": False,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+}
+
+SCATACSEQ_LOWER_VERSION_VALID = {
+    "test-schema-v0.0": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": False,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+    "test-schema-v0.1": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": True,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+}
+
+SCATACSEQ_NEITHER_VERSION_VALID = {
+    "test-schema-v0.0": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": True,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+    "test-schema-v0.1": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": True,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+}
+
+SCATACSEQ_BOTH_VERSIONS_VALID = {
+    "test-schema-v0.0": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": False,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+    "test-schema-v0.1": {
+        "files": [
+            {
+                "pattern": "[^/]+\\.fastq\\.gz",
+                "description": "Compressed FastQ file",
+                "required": True,
+            },
+            {
+                "pattern": "extras\\/.*",
+                "required": False,
+                "description": "Folder for general lab-specific files related to the dataset. [Exists in all assays]",
+            },
+        ]
+    },
+}